refactor(settings): move users into team access settings
