+++ frontend/src/app/settings/(team)/team/page.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/components/users/UsersManagementPage";
